+++ R/estimate_infections.R
@@ -1,11 +1,14 @@
 #' Estimate Infections, the Time-Varying Reproduction Number and the Rate of Growth
 #'
 #' @description This function uses a non-parametric approach to reconstruct cases by date of infection from reported 
-#' cases. It can optionally then estimate the time-varying reproduction number and the rate of growth.
+#' cases. This function uses either a generative Rt model or non-parametric back calculation to estimate underlying
+#' latent infections and then maps these infections to observed cases via uncertain reporting delays and a flexible
+#' observation model. See the examples and function arguments for the details of all options. The default settings
+#'  may not be sufficient for your use case so the number of warmup samples (`stan_args = list(warmup)`) may need to
+#'  be increased as may the overall number of samples. Follow the links provided by any warnings messages to diagnose 
+#'  issues with the MCMC fit.
 #' @param reported_cases A data frame of confirmed cases (confirm) by date (date). confirm must be integer and date must be 
 #' in date format.
-#' @param family A character string indicating the reporting model to use. Defaults to negative 
-#' binomial ("negbin") with poisson ("poisson") also supported.
 #' @param generation_time A list containing the mean, standard deviation of the mean (mean_sd), 
 #' standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
 #' generation time (assuming a gamma distribution).
@@ -14,38 +17,21 @@
 #' standard deviation (sd), standard deviation of the standard deviation and the maximum allowed value for the
 #' that delay (assuming a lognormal distribution with all parameters excepting the max allowed value 
 #' on the log scale). To use no delays set this to `list()`.
-#' @param rt_prior A list contain the mean and standard deviation (sd) of the lognormally distributed prior for
-#' Rt. By default this is assumed to be mean 1 with a standard deviation of 1 (note in model these will be mapped to
+#' @param rt_prior A list contain the mean and standard deviation (sd) of the lognormal prior for Rt. By default this is assumed to be mean 1 with a standard deviation of 1 (note in model these will be mapped to
 #' log space). To infer infections and then calculate Rt using backcalculation set this to `list()`.
 #' @param prior_smoothing_window Numeric defaults to 7. The number of days over which to take a rolling average
 #' for the prior based on reported cases.
 #' @param horizon Numeric, defaults to 7. Number of days into the future to forecast.
 #' @param model A compiled stan model. By default uses the internal package model.
 #' @param samples Numeric, defaults to 1000. Number of samples post warmup.
-#' @param week_effect Logical, defaults TRUE. Should weekly reporting effects be estimated.
 #' @param use_breakpoints Logical, defaults to TRUE but only active if a `breakpoint` variable is present in the input data. 
 #'  Breakpoints should be defined as 1 if present and otherwise 0. By default breakpoints are fit jointly with
 #' a global non-parametric effect and so represent a conservative estimate of breakpoint changes. To specify a random walk define
 #' breakpoints every n days (so every 7 days for a weekly random walk) and disable the gaussian process using `gp = list()`.
-#' @param burn_in Numeric, defaults to 0. The number of initial Rt estimates to discard. This argument may be used to reduce 
-#' spurious findings when running `estimate_infections` on a partial timeseries (as the earliest estimates will not be informed by 
-#' all cases that occurred only those supplied to `estimate_infections`). The combined delays used will inform the appropriate length
-#' of this burn in but 7 days is likely a sensible starting point.
-#' @param stationary Logical, defaults to FALSE. Should Rt be estimated with a global mean. When estimating Rt 
-#' this should substantially improve run times but will revert to the global average for real time and forecasted estimates.
-#' This setting is most appropriate when estimating historic Rt or when combined with breakpoints.
-#' @param return_fit Logical, defaults to FALSE. Should the fitted stan model be returned.
-#' @param gp List controlling the Gaussian process approximation if set to `list()` then `Rt` is assumed to be constant unless
-#' other settings introduce variation. If set must contain the `basis_prop` (number of basis functions based on scaling the time points)
-#'  which defaults to 0.3 and must be between 0 and 1 (increasing this increases the accuracy of the approximation and the cost of 
-#'  additional compute. Must also contain the `boundary_scale` (multiplied by half the range of the input time series). Increasing this 
-#' increases the accuracy of the approximation at the cost of additional compute. 
-#' See here: https://arxiv.org/abs/2004.11408 for more information on setting these parameters.
-#' Must also contain the  `lengthscale_alpha` and `lengthscale_beta`. These tune the prior of the lengthscale. Principled 
-#' values can be obtained using `tune_inv_gamma` which optimises based on the desired truncation (which should be based on the scale
-#' of the observed data). The default is tuned to have 98% of the density of the distribution between 2 and 21 days. Finally the list must 
-#' contain `alpha_sd` the standard deviation for the alpha parameter of the gaussian process. This defaults to 0.1.
-#' @param verbose Logical, defaults to `FALSE`. Should verbose debug progress messages be printed. Corresponds to the "DEBUG" level from 
+#' @param burn_in Numeric, defaults to 0. The number of initial Rt estimates to discard. This argument is depreciated and will be 
+#' removed from the package unless a clear user need is given.
+#' @param return_fit Logical, defaults to TRUE. Should the fitted stan model be returned.
+#' @param verbose Logical, defaults to `TRUE` when used interactively and otherwise `FALSE`. Should verbose debug progress messages be printed. Corresponds to the "DEBUG" level from 
 #' `futile.logger`. See `setup_logging` for more detailed logging options.
 #' @param future Logical, defaults to `FALSE`. Should stan chains be run in parallel using `future`. This allows users to have chains
 #' fail gracefully (i.e when combined with `max_execution_time`). Should be combined with a call to `future::plan`
@@ -54,7 +40,8 @@
 #' will fail with an informative error.
 #' @export
 #' @inheritParams create_stan_args
-#' @inheritParams create_future_rt
+#' @inheritParams create_stan_data
+#' @inheritParams create_gp_data
 #' @inheritParams fit_model_with_nuts
 #' @inheritParams calc_CrIs
 #' @importFrom data.table data.table copy merge.data.table as.data.table setorder rbindlist setDTthreads melt .N setDT
@@ -65,57 +52,38 @@
 #' @examples
 #' \donttest{
 #' # get example case counts
-#' reported_cases <- EpiNow2::example_confirmed[1:50]
+#' reported_cases <- EpiNow2::example_confirmed[1:60]
 #' 
 #' # set up example generation time
 #' generation_time <- get_generation_time(disease = "SARS-CoV-2", source = "ganyani")
 #' # set delays between infection and case report 
 #' incubation_period <- get_incubation_period(disease = "SARS-CoV-2", source = "lauer")
-#' reporting_delay <- list(mean = log(3), mean_sd = 0.1,
-#'                         sd = log(1), sd_sd = 0.1, max = 15)
+#' reporting_delay <- list(mean = convert_to_logmean(3, 1), 
+#'                         mean_sd = 0.1,
+#'                         sd = convert_to_logsd(3, 1), 
+#'                         sd_sd = 0.1, 
+#'                         max = 15)
 #'       
-#' # Note: all examples below have been tuned to reduce the runtimes of examples
-#' # these settings are not suggesed for real world use.                   
 #' # run model with default setting
 #' def <- estimate_infections(reported_cases, generation_time = generation_time,
-#'                            delays = list(incubation_period, reporting_delay), 
-#'                            stan_args = list(warmup = 200,
-#'                             control = list(adapt_delta = 0.95, max_treedepth = 15),
-#'                             cores = ifelse(interactive(), 4, 1)),
-#'                             verbose = interactive())
-#'
+#'                            delays = list(incubation_period, reporting_delay),
+#'                            stan_args = list(cores = ifelse(interactive(), 4, 1)))
 #' plot(def)
 #' 
-#' # run model using backcalculation
+#' # run model using backcalculation (combined here with under reporting)
 #' backcalc <- estimate_infections(reported_cases, generation_time = generation_time,
 #'                                 delays = list(incubation_period, reporting_delay),
-#'                                 stan_args = list(warmup = 200, 
-#'                                                  cores = ifelse(interactive(), 4, 1),
-#'                                                  control = list(adapt_delta = 0.9)),
-#'                                 rt_prior = list(), verbose = interactive())
-#'
+#'                                 rt_prior = list(),
+#'                                 obs_model = list(scale = list(mean = 0.4, sd = 0.01)),
+#'                                 stan_args = list(cores = ifelse(interactive(), 4, 1)))
 #' plot(backcalc)
-#' 
-#' # run the model with default settings using the future backend 
-#' ## (combine with a call to future::plan to make this parallel).
-#' def_future <- estimate_infections(reported_cases, generation_time = generation_time,
-#'                                   delays = list(incubation_period, reporting_delay),
-#'                                   stan_args = list(warmup = 200, 
-#'                                                    control = list(adapt_delta = 0.9),
-#'                                                    cores = ifelse(interactive(), 4, 1)),
-#'                                   verbose = interactive(), future = TRUE)
-#' 
-#' plot(def_future)                        
 #'                            
-#' # run model with Rt fixed into the future using the latest estimate
-#' fixed_rt <- estimate_infections(reported_cases, generation_time = generation_time,
-#'                                 delays = list(incubation_period, reporting_delay),
-#'                                 stan_args = list(warmup = 200, 
-#'                                                  control = list(adapt_delta = 0.9),
-#'                                                  cores = ifelse(interactive(), 4, 1)),
-#'                                 future_rt = "latest", verbose = interactive())
-#'
-#' plot(fixed_rt)
+#' # run model with Rt projected into the future using the Gaussian process
+#' project_rt <- estimate_infections(reported_cases, generation_time = generation_time,
+#'                                   delays = list(incubation_period, reporting_delay),
+#'                                   gp = list(future = "project"),
+#'                                   stan_args = list(cores = ifelse(interactive(), 4, 1)))
+#' plot(project_rt)
 #'
 #' # run the model with default settings on a later snapshot of 
 #' # data (use burn_in here to remove the first week of estimates that may
@@ -123,84 +91,74 @@
 #' snapshot_cases <- EpiNow2::example_confirmed[80:130]
 #' snapshot <- estimate_infections(snapshot_cases, generation_time = generation_time,
 #'                                 delays = list(incubation_period, reporting_delay),
-#'                                 stan_args = list(warmup = 200, 
-#'                                                  control = list(adapt_delta = 0.9),
-#'                                                  cores = ifelse(interactive(), 4, 1)),
-#'                                 burn_in = 7, verbose = interactive())
-#'
+#'                                 stan_args = list(cores = ifelse(interactive(), 4, 1)))
 #' plot(snapshot) 
 #' 
 #' # run model with stationary Rt assumption (likely to provide biased real-time estimates)
 #' stat <- estimate_infections(reported_cases, generation_time = generation_time,
 #'                             delays = list(incubation_period, reporting_delay),
-#'                             stan_args = list(warmup = 200, cores = ifelse(interactive(), 4, 1),
-#'                                              control = list(adapt_delta = 0.9)),
-#'                             stationary = TRUE, verbose = interactive())
-#'
+#'                             gp = list(stationary = TRUE),
+#'                             stan_args = list(cores = ifelse(interactive(), 4, 1)))
 #' plot(stat)
 #'        
-#' # run model with fixed Rt assumption 
+#' # run model without a gaussian process (i.e fixed Rt assuming no breakpoints)
 #' fixed <- estimate_infections(reported_cases, generation_time = generation_time,
 #'                              delays = list(incubation_period, reporting_delay),
-#'                              stan_args = list(warmup = 200, cores = ifelse(interactive(), 4, 1),
-#'                                               control = list(adapt_delta = 0.9)),
-#'                              gp = list(), verbose = interactive())
-#'
+#'                              gp = NULL,
+#'                              stan_args = list(cores = ifelse(interactive(), 4, 1)))
 #' plot(fixed)
 #' 
 #' # run model with no delays 
 #' no_delay <- estimate_infections(reported_cases, generation_time = generation_time,
-#'                                 stan_args = list(warmup = 200,
-#'                                                  cores = ifelse(interactive(), 4, 1),
-#'                                                  control = list(adapt_delta = 0.9)),
-#'                                 verbose = interactive())
-#'
+#'                                 stan_args = 
+#'                                      list(warmup = 200,
+#'                                           cores = ifelse(interactive(), 4, 1),
+#'                                           control = list(adapt_delta = 0.95, max_treedepth = 15)))
 #' plot(no_delay)    
-#'      
-#' # add a dummy breakpoint (used only when optionally estimating breakpoints)
-#' reported_cases_bp <- data.table::copy(reported_cases)[,
-#'               breakpoint := data.table::fifelse(date == as.Date("2020-03-16"), 1, 0)]              
+#' 
 #' # run model with breakpoints but otherwise static Rt
-#' # This formulation may increase the apparent effect of the breakpoint but needs to be tested using
-#' # model fit criteria (i.e LFO), comparison to models with non-static Rt and models with partially
-#' # constrained non-static Rt.           
+#' # add a dummy breakpoint (used only when optionally estimating breakpoints)
+#' reported_cases_bp <- 
+#'   data.table::copy(reported_cases)[, breakpoint := ifelse(date == as.Date("2020-03-16"), 
+#'                                                           1, 0)]
 #' bkp <- estimate_infections(reported_cases_bp, generation_time = generation_time,
-#'                             delays = list(incubation_period, reporting_delay),
-#'                             stan_args = list(warmup = 200, 
-#'                                              cores = ifelse(interactive(), 4, 1),
-#'                                              control = list(adapt_delta = 0.9)),
-#'                             gp = list(), verbose = interactive())                                                         
-#'
+#'                            delays = list(incubation_period, reporting_delay),
+#'                            stan_args = 
+#'                               list(warmup = 200, 
+#'                                    cores = ifelse(interactive(), 4, 1),
+#'                                    control = list(adapt_delta = 0.95, max_treedepth = 15)),
+#'                            gp = NULL)                                                         
 #' plot(bkp)
 #' # breakpoint effect
-#' fbkp$summarised[variable == "breakpoints"]
+#' bkp$summarised[variable == "breakpoints"]
 #' }                                
 estimate_infections <- function(reported_cases, 
+                                generation_time, 
+                                delays = list(),
+                                obs_model = list(),
+                                stan_args = list(),
+                                gp = list(),
+                                rt_prior = list(mean = 1, sd = 0.5),
+                                horizon = 7,
                                 model = NULL, 
                                 samples = 1000,
-                                stan_args = NULL,
                                 method = "exact", 
-                                family = "negbin", 
-                                generation_time, 
-                                CrIs = c(0.2, 0.5, 0.9),
-                                delays = list(),
-                                horizon = 7,
-                                gp = list(basis_prop = 0.3, boundary_scale = 2, 
-                                          lengthscale_alpha = 4.5, lengthscale_beta = 21.5,
-                                          alpha_sd = 0.1),
-                                rt_prior = list(mean = 1, sd = 0.5),
-                                week_effect = TRUE, 
                                 use_breakpoints = TRUE, 
-                                stationary = FALSE, 
-                                future_rt = "project",
                                 burn_in = 0, 
                                 prior_smoothing_window = 7, 
+                                CrIs = c(0.2, 0.5, 0.9),
                                 future = FALSE, 
                                 max_execution_time = Inf, 
-                                return_fit = FALSE,
+                                return_fit = TRUE,
                                 id = "estimate_infections",
-                                verbose = FALSE){
+                                verbose = interactive()){
    
+  if (burn_in > 0) {
+    futile.logger::flog.info("burn_in is depreciated as of EpiNow2 1.3.0 - if using 
+                             this feature please contact the developers",
+                             name = "EpiNow2.epinow.estimate_infections")
+  }
+  
   # store dirty reported case data
   dirty_reported_cases <- data.table::copy(reported_cases)
   # set fall back rt prior and trigger switches
@@ -211,18 +169,6 @@ estimate_infections <- function(reported_cases,
     estimate_rt <- TRUE
   }
   
-  # If no GP default to stationary setup with no assumed non-parametric change
-  # add in placeholder parameters
-  if (length(gp) == 0) {
-    fixed <- TRUE
-    stationary <- TRUE
-    gp = list(basis_prop = 1, boundary_scale = 1,
-              lengthscale_alpha = 1, lengthscale_beta = 1,
-              alpha_sd = 0.1)
-  }else{
-    fixed <- FALSE
-  }
-
   # Check verbose settings and set logger to match---------------------------
   if (verbose) {
     futile.logger::flog.threshold(futile.logger::DEBUG,
@@ -247,7 +193,6 @@ estimate_infections <- function(reported_cases,
  
   # Organise delays ---------------------------------------------------------
   no_delays <- length(delays)
-  
   if (no_delays > 0) {
     delays <- purrr::transpose(delays)
   }
@@ -293,13 +238,10 @@ estimate_infections <- function(reported_cases,
                            generation_time = generation_time,
                            rt_prior = rt_prior, 
                            estimate_rt = estimate_rt,
-                           week_effect = week_effect, 
-                           stationary = stationary,
-                           fixed = fixed,
+                           burn_in = burn_in,
                            break_no = break_no, 
-                           future_rt = future_rt, 
                            gp = gp,
-                           family = family,
+                           obs_model = obs_model,
                            delays = delays)
 
   # Set up default settings -------------------------------------------------
@@ -344,6 +286,7 @@ estimate_infections <- function(reported_cases,
   ## Join stan fit if required
   if (return_fit) {
     format_out$fit <- fit
+    format_out$args <- data
   }
   format_out$observations <- dirty_reported_cases
   class(format_out) <- c("estimate_infections", class(format_out))
@@ -520,6 +463,7 @@ fit_model_with_vb <- function(args, future = FALSE, id = "stan", verbose = FALSE
 #' @inheritParams calc_summary_measures
 #' @importFrom data.table fifelse rbindlist 
 #' @importFrom lubridate days
+#' @importFrom futile.logger flog.info
 #' @return A list of samples and summarised posterior parameter estimates
 format_fit <- function(posterior_samples, horizon, shift, burn_in, start_date,
                        CrIs){
@@ -540,6 +484,9 @@ format_fit <- function(posterior_samples, horizon, shift, burn_in, start_date,
   
   # remove burn in period if specified
   if (burn_in > 0) {
+    futile.logger::flog.info("burn_in is depreciated as of EpiNow2 1.3.0 - if using 
+                             this feature please contact the developers",
+                             name = "EpiNow2.epinow.estimate_infections")
     format_out$samples <- format_out$samples[is.na(date) | date >= (start_date + lubridate::days(burn_in))]
   }