+++ src/RcppExports.cpp
@@ -11,6 +11,7 @@ RcppExport SEXP _rcpp_module_boot_stan_fit4estimate_infections_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4exp_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4gamma_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4lnorm_mod();
+RcppExport SEXP _rcpp_module_boot_stan_fit4simulate_infections_mod();
 RcppExport SEXP _rcpp_module_boot_stan_fit4tune_inv_gamma_mod();
 
 static const R_CallMethodDef CallEntries[] = {
@@ -18,6 +19,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_rcpp_module_boot_stan_fit4exp_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4exp_mod, 0},
     {"_rcpp_module_boot_stan_fit4gamma_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4gamma_mod, 0},
     {"_rcpp_module_boot_stan_fit4lnorm_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4lnorm_mod, 0},
+    {"_rcpp_module_boot_stan_fit4simulate_infections_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4simulate_infections_mod, 0},
     {"_rcpp_module_boot_stan_fit4tune_inv_gamma_mod", (DL_FUNC) &_rcpp_module_boot_stan_fit4tune_inv_gamma_mod, 0},
     {NULL, NULL, 0}
 };